- Added PEThreatIntelligence header and implementation for threat intelligence integration - Refactored peCommon into separate header and source files - Enhanced OutputManager and PESuspiciousTechniqueAnalyzer for improved reporting - Updated Makefile, .gitignore, and README.md to reflect new modules and usage - Improved logging and resource parsing in Logger and PEResourceParser - General code cleanup and modularization

+++ src/PEResourceParser.cpp
@@ -434,7 +434,7 @@ std::vector<IconInfo> PEResourceParser::parseIconResource(const std::vector<uint
         return icons;
     }
 
-    for (uint16_t i = 0; i < iconCount && (6 + i * 16) < data.size(); i++) {
+    for (uint16_t i = 0; i < iconCount && (6 + static_cast<size_t>(i) * 16) < data.size(); i++) {
         IconInfo iconInfo = parseIconEntry(dataPtr + 6 + i * 16, 16);
         if (iconInfo.isValid) {
             icons.push_back(iconInfo);
